Move writechecker apis to api.go
